requirements.txt
requests
pytest
Pillow
numpy